\name{act_29pt}
\alias{act_29pt}
\docType{data}
\title{Data Set Containing a One Day Average of Actigraph Data for 29 Subjects}
\description{A data set containing a one day(1440 minutes) of actigraph data for 29 subjects.}

\usage{data(act_29pt)}

\format{
	A data frame in which each column contains the actigraph data of a subject with 
	column names being the subject IDs. The row names are the time indices of one 
	day by minute from midnight to midnight. 
}

\keyword{datasets}