\name{act_8pt}
\docType{data}
\alias{act_8pt}
\title{Data Set Containing a One Day Average of Actigraph Data for 8 Subjects}
\description{A data set containing a one day(1440 minutes) of actigraph data for 8 subjects.}

\usage{data(act_8pt)}

\format{
	A data frame that consists of nine columns; the first column refers to the time
	indices of one day by minute from midnight to midnight. The following eight columns contain 
	the actigraph data of eight subjects with column names being the subject IDs.
}

\keyword{datasets}