\name{clinic_8pt}
\docType{data}
\alias{clinic_8pt}
\title{Data Set Containing AHI(Apnea Hypopnea Index) Information from 8 Subjects}
\description{A data set containing the AHI(Apnea Hypopnea Index) values(categorical) for 8 subjects.}

\usage{data(clinic_8pt)}

\format{
	A data frame that consists of two columns. The first contains the numeric IDs 
	and the second column contains the AHI information for 8 subjects.
}

\keyword{datasets}