% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crfpca.R
\name{crfpca}
\alias{crfpca}
\title{Functional Principal Component Analysis for Whole Day Circadian Rhythmicity}
\usage{
crfpca(count.data, logtransform = FALSE, knots = 20, pve = 0.9)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n*1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{logtransform}{Conduct log transfomation before FPCA. Default is false.}

\item{knots}{number of knots to use or the vectors of knots; defaults to 20.}

\item{pve}{proportion of variation explained; defaults to 0.9.}
}
\value{
Alist with elements
\item{phi}{Principal component loadings, aka eigen functions}
\item{pcs}{Principal component scores}
}
\description{
A nonparametric approach to study variational stucture of whole day
activity trajectory. This function provides the principal component loading functions and
corresponding principal component scores. FPCA is done via the sandwich smoother for covriance matrix smoothing.=
}
\examples{

res = crfpca(count.data = count, logtransform  = TRUE)


}
