% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_activity_data.R
\docType{data}
\name{example_activity_data}
\alias{example_activity_data}
\title{Example Activity/Wear Data from Multiple People}
\format{A list of two \code{data.frame}s with 1442 columns, which are in the following order:
\describe{
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{MIN1-MIN1440}{counts of activity}
}}
\usage{
example_activity_data
}
\description{
A list of two data.frames containing the counts and the weartime
for multiple participants, with a column for the ID and a column
for the visit.
}
\keyword{datasets}
