% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtCos.R
\name{ExtCos}
\alias{ExtCos}
\title{Extended Cosinor Model for Circadian Rhythmicity}
\usage{
ExtCos(x, logtransform = TRUE)
}
\arguments{
\item{x}{\code{vector} vector of dimension n*1440 which reprsents n days of 1440 minute activity data}

\item{longtransform}{Conduct log transfomation before. Default is \code{TRUE}.}
}
\value{
A list with elements
\item{min}{minimum}
\item{amp}{amplitude}
\item{alpha}{alpha parameter}
\item{beta}{beta parameter}
\item{acro}{acrophase}
}
\description{
A parametric approach to study circadian rhythmicity assuming cosinor shape.
Here we used the anti logistic transformed cosine curves, which provided 5 parameters.
}
\examples{
coutn.day = rpois(1440*5, lambda = 5)
extcos = ExtCos(x = coutn.day, logtransform  = TRUE)


}
