% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.R
\name{IV}
\alias{IV}
\title{Intradaily Variability}
\usage{
IV(x, level = c("minute", "hour"))
}
\arguments{
\item{x}{\code{vector} vector of dimension 440 which reprsents 1440 minute activity data}

\item{level}{time resolution to calcualte IV. Can be either "minute", or "hour".}
}
\value{
IV
}
\description{
This function calcualte intradaily variability, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
iv = IV(x = count1, level = "hour")


}
