% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_long.R
\name{IV_long}
\alias{IV_long}
\title{Intradaily Variability for Whole Dataset}
\usage{
IV_long(count.data, level = c("minute", "hour"))
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n * 1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{level}{time resolution to calcualte IV. Can be either "minute", or "hour".}
}
\value{
A \code{data.frame} with following columns
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{IV_min}{minute level IV}
\item{IV_hour}{hourly level IV parameter}
}
\description{
Intradaily variability is a nonparametric metric
reprsenting fragmentation of circadian rhtymicity. this function is a whole data wrapper for
\code{IV}.
}
\examples{
data(example_activity_data)
count.data = example_activity_data$count
iv_all = IV_long(count.data = count.data, level = "hour")



}
