% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot Daily Activity Profile}
\usage{
plot_profile(x, w, title = NULL, cex.main = 1, cex.sub = 2, cex.lab = 1,
  hr = 3, at.y = NULL, cex.xaxis = 1, cex.yaxis = 1, oma = c(1, 1, 1,
  1), mar = c(4, 4, 2, 1), ylab = "AC")
}
\arguments{
\item{x}{\code{vector} of activity data.}

\item{w}{\code{vector} of wear flag data with same dimension as \code{x}.}

\item{title}{title or the plot}

\item{cex.main}{size of title}

\item{cex.sub}{size of subtitle}

\item{cex.lab}{size of x and y lable}

\item{hr}{how many hours away between two x axis annotation ticks,
has to be \code{integer}; defualts to be 3.}

\item{at.y}{where to put y axis annotation (ticks)}

\item{cex.xaxis}{size of x axis annotation}

\item{cex.yaxis}{size of y axis annotation}

\item{oma}{defaults to c(1, 1, 1, 1)}

\item{mar}{defaults to c(4, 4, 2, 1)}

\item{ylab}{y axis label}
}
\description{
Given a activity vector (and a wear/nonwear flag vector), plot the
daily activity profile (with background)
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
wear1 = c(t(example_activity_data$wear[1,-c(1,2)]))
plot_profile(x=count1, w=wear1, title = "Example")


}
