% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA.R
\name{RA}
\alias{RA}
\title{Relative Amplitude}
\usage{
RA(x)
}
\arguments{
\item{x}{\code{vector} vector of dimension 440 which reprsents 1440 minute activity data}
}
\value{
RA
}
\description{
This function calcualte relative amplitude, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
ra = RA(x = count1)


}
