% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA_long.R
\name{RA_long}
\alias{RA_long}
\title{Relative Amplitude for Whole Dataset}
\usage{
RA_long(count.data)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n * 1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}
}
\value{
A \code{data.frame} with following columns
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{RA}{RA}

#' @importFrom zoo rollapplyr
}
\description{
Relative Amplitude is a nonparametric metric
reprsenting fragmentation of circadian rhtymicity. this function is a whole data wrapper for
\code{RA}.
}
\examples{
data(example_activity_data)
count.data = example_activity_data$count
ra_all = RA_long(count.data = count.data)



}
