% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R
\name{Time}
\alias{Time}
\title{Time of A Certian State}
\usage{
Time(x, w, thresh, smallerthan = TRUE, bout.length = 1)
}
\arguments{
\item{x}{\code{vector} of activity data.}

\item{w}{\code{vector} of wear flag data with same dimension as \code{x}.}

\item{thresh}{threshold to binarize the data.}

\item{smallerthan}{Find a state that is smaller than a threshold, or greater than or equal to.}

\item{bout.length}{minimum duration of defining an active bout; defaults to 1.}
}
\value{
Time
}
\description{
Calculate the total time of being in certain state, e.g. sedentary, active, MVPA, etc.
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
wear1 = c(t(example_activity_data$wear[1,-c(1,2)]))
time_sedentary = Time(x = count1, w = wear1, thresh = 100,
bout.length = 1, smallerthan = TRUE)


}
