% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tvol.R
\name{Tvol}
\alias{Tvol}
\title{Total Volumen of Activity for Whole Dataset}
\usage{
Tvol(count.data, weartime, logtransform = FALSE)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n*1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{weartime}{\code{data.frame} with dimension of \code{count.data}.
The first two columns have to be ID and Day.}

\item{logtransform}{if \code{TRUE}, then calcualte \code{TLAC}. Or calculate \code{TAC}.}
}
\value{
A dataframe with some of the following columns
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{TAC}{total activity count}
\item{TLAC}{total log activity count}
}
\description{
Calculate total volume of activity level, which includes
\code{TLAC} (total log transfored activity counts),
\code{TAC} (total activity counts).
}
\details{
log transormation is defined as log(x+1).
}
\examples{
data(example_activity_data)
count = example_activity_data$count
wear = example_activity_data$wear
tac = Tvol(count.data = count,weartime = wear,logtransform = FALSE)

}
