\name{weekday}
\docType{data}
\alias{weekday}
\title{Data Set Containing Five Weekdays of Actigraph Data for One Subject}
\description{A data set containing five weekdays of actigraph data for one subject.}

\usage{data(weekday)}

\format{
	A data frame that consists of three columns; the first column is the weekday 
	indices, the second column contains the indices of time in minutes from midnight to 
	midnight, and the third contains the actigraph data.
}

\keyword{datasets}