\name{Actigraphy-package}
\alias{Actigraphy-package}
\alias{Actigraphy}
\docType{package}
\title{Functional Actigraphy Data Analysis}
\description{This package implements functional linear modeling and analysis for actigraphy data.}
\details{
	\tabular{ll}{
		Package: \tab Actigraphy\cr
		Type: \tab Package\cr
		Version: \tab 1.1\cr
		Date: \tab 2012-09-19\cr
		License: \tab GPL-2\cr
		LazyLoad: \tab yes\cr
	}
	
	For a complete list of functions with individual help pages, use library(help="Actigraphy"). 
	
	Please refer to the directory "Actigraphy/doc" for an additional tutorials \cr
	and a document containing the code for the figures in the referenced paper. \cr
	
	The paper can be downloaded from the Journal of Circadian Rhythms website. \cr
}
\author{
William Shannon, Tao Li, Hong Xian, Jia Wang, Elena Deych, Carlos Gonzalez \cr
Maintainer: Berkley Shands <bshands@dom.wustl.edu>
}
\references{1. "Measuring the Impact of AHI and Obesity on Circadian Activity Patterns 
Using Functional Linear Modeling of Actigraphy Data," Jia Wang, Amy Licis, 
Elena Deych, Jimin Ding, Jennifer McLeland, Cristina Toedebusch, Tao Li, 
Hong Xian, Stephen Duntley, and William Shannon.
}

\keyword{ package }