% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot Daily Activity Profile}
\usage{
plot_profile(x, w, title = NULL, cex.main = 1, cex.sub = 2, cex.lab = 1,
  at.y = NULL, cex.xaxis = 1.5, cex.yaxis = 1.5)
}
\arguments{
\item{x}{\code{vector} of activity data.}

\item{w}{\code{vector} of wear flag data with same dimension as \code{x}.}

\item{title}{title or the plot}

\item{cex.main}{size of title}

\item{cex.sub}{size of subtitle}

\item{cex.lab}{size of x and y lable}

\item{at.y}{where to put y axis annotation}

\item{cex.xaxis}{size of x axis annotation}

\item{cex.yaxis}{size of y axis annotation}

\item{metrics}{What is the fragmentation metrics to exract. Can be
"mean_bout","TP","Gini","power","hazard",or all the above metrics "all".}
}
\value{
A list with elements
\item{mean_r}{mean sedentary bout duration}
\item{mean_a}{mean active bout duration}
\item{SATP}{sedentary to active transition probability}
\item{ASTP}{bactive to sedentary transition probability}
\item{Gini_r}{Gini index for active bout}
\item{Gini_a}{Gini index for sedentary bout}
\item{h_r}{hazard function for sedentary bout}
\item{h_a}{hazard function for active bout}
\item{alpha_r}{power law parameter for sedentary bout}
\item{alpha_a}{power law parameter for active bout}
}
\description{
Given a activity vector (and a wear/nonwear flag vector), plot the
daily activity profile (with background)
}
\examples{
data(example_activity_data)
count1 = c(t(example_activity_data$count[1,-c(1,2)]))
wear1 = c(t(example_activity_data$wear[1,-c(1,2)]))
plot_profile(x=count1, w=wear1, title = "Example")


}
